chore: copy unity built files after publish-preview (#4351)
* chore: move publish-artifacts after we publish the package

* rename publish-artifacts to publish-preview

* separate copyBuiltFiles from publish process

* fix
