+++ browser-interface/package.json
@@ -9,7 +9,8 @@
     "test": "make test-ci",
     "publish": "ts-node --project scripts/tsconfig.json scripts/publish.ts",
     "check-file-size": "ts-node --project scripts/tsconfig.json scripts/checkFileSizes.ts",
-    "publish-artifacts": "./scripts/publish-artifacts.sh"
+    "copy-built-files": "ts-node --project scripts/tsconfig.json scripts/copyBuiltFiles.ts",
+    "publish-preview": "./scripts/publish-preview-artifacts.sh"
   },
   "generatedFiles": {
     "frameworkUrl": "unity.framework.js",
